+++ src/main/kotlin/com/arkhamusserver/arkhamus/view/dto/netty/response/ritual/GodVoteStartNettyResponse.kt
@@ -5,7 +5,7 @@ import com.arkhamusserver.arkhamus.view.dto.netty.response.*
 import com.arkhamusserver.arkhamus.view.dto.netty.response.parts.*
 
 class GodVoteStartNettyResponse(
-    private val godId: Long?,
+    private val godId: Int?,
     private val executedSuccessfully: Boolean,
     private val firstTime: Boolean,
     tick: Long,
